+++ lib/js/jsx.js
@@ -1,19 +1,1 @@
-'use strict';
-
-var Curry = require("./curry.js");
-
-function createElementWithKey(key, createElement, component, props) {
-  return Curry._2(createElement, component, key !== undefined ? Object.assign({}, props, {
-                    key: key
-                  }) : props);
-}
-
-function createElementVariadicWithKey(key, createElementVariadic, component, props, elements) {
-  return Curry._3(createElementVariadic, component, key !== undefined ? Object.assign({}, props, {
-                    key: key
-                  }) : props, elements);
-}
-
-exports.createElementWithKey = createElementWithKey;
-exports.createElementVariadicWithKey = createElementVariadicWithKey;
-/* No side effect */
+/* This output is empty. Its source's type definitions, externals and/or unused code got optimized away. */
